refactor readme and file structure
