+++ README.md
@@ -1,3 +1,5 @@
 # DataCommProj1
 
-These scripts should be in different directories to confirm they are working. 
+This project will demonstrate downloading and uploading files through an FTP server and a client. This is the first project for CIS457.
+
+Created by: Julia Hildebrand, AJ Natzic, Sia Niotis and Alex Porter
