+++ client/SomethingToUpload.txt
@@ -1,0 +1,2 @@
+This is some text that you can upload to the server.
+It is not very long.
