+++ server/SomethingToDownload.txt
@@ -1,0 +1,2 @@
+This is something that you can download.
+It is also pretty small.
